Atualização da dependencia Repository para Renda

+++ apimobfin/src/main/java/com/tiraiz/apimobfin/repository/RendaRepository.java
@@ -4,6 +4,6 @@ import com.tiraiz.apimobfin.models.Renda;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RendaRepository extends JpaRepository<Renda, Long> {
-    Renda findByid(int id);
+    Renda findByid(long id);
 
 }
